fix: Pass the correct _rev to account deletion job
